# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: American Auto Parts Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles/models:
    get:
      operationId: getVehicleModels
      tags: [vehicles]
      summary: Get models for a make
      parameters:
        - name: make
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Vehicle models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VehicleModel"
  /vehicles/years:
    get:
      operationId: getVehicleYears
      tags: [vehicles]
      summary: Get available years for a make/model
      parameters:
        - name: make
          in: query
          required: true
          schema: { type: string }
        - name: model
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Years
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer

  # ─── Wishlist ─────────────────────────────────────────────────────────────
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Wishlist updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Wishlist updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ─── Users ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        sku: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        imageUrl: { type: string }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        categoryName: { type: string }
        brandId: { type: integer }
        brandName: { type: string }
        stock: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        compatibility: { type: array, items: { type: string } }
        specs:
          $ref: "#/components/schemas/ProductSpecs"
        warranty: { type: ["string", "null"] }
        manufacturer: { type: ["string", "null"] }
        isFeatured: { type: boolean }
        createdAt: { type: string }
      required:
        [id, sku, name, description, price, imageUrl, images, categoryId, categoryName, brandId, brandName, stock, rating, reviewCount, compatibility, isFeatured, createdAt]
...
    VehicleModel:
      type: object
      properties:
        model: { type: string }
        years: { type: array, items: { type: integer } }
      required: [model, years]

    # ─── Users ───────────────────────────────────────────────────────────────
...
    ProductSpecs:
      type: object
      properties:
        material: { type: string }
        weight: { type: string }
        dimensions: { type: string }
        partNumber: { type: string }
        fitment: { type: string }
      additionalProperties: false
...