# lib/api-zod/src/generated/types/vehicleModel.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * American Auto Parts Store API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleModel {
  model: string;
  years: number[];
}